+++ .github/workflows/ci.yml
@@ -421,7 +421,7 @@ jobs:
           use-sticky-disk: "false"
 
       - name: Build dist
-        run: pnpm build:ci-artifacts
+        run: pnpm build
 
       - name: Build Control UI
         run: pnpm ui:build
