+++ docs/.generated/config-baseline.sha256
@@ -1,4 +1,4 @@
-b2722c773d6e3bb3fb6c2492a8b710e2ff646b8db44e595abfdba61e67c965ed  config-baseline.json
-2dac6cea7a03051e10c4f1ceeb8025c23220a550b940156d6eddeb8e3ecd75f1  config-baseline.core.json
+3583489dfebd88a53f1c66c984b16dc5eff752c887d4c582a86753990f1d5b18  config-baseline.json
+a490b20c47a45c3e26b6917eb3e102356698395128aec20b1f4aabb62ca7cad1  config-baseline.core.json
 3bb312dc9c39a374ca92613abf21606c25dc571287a3941dac71ff57b2b5c519  config-baseline.channel.json
-d8a898c54e90555dac185eda9dfc1f18e37bb09e4076c7cf3c7ed9ac88f3a67b  config-baseline.plugin.json
+0471a5bffb213a3829555efe5961f5b5fd5080c1d38b1ac8dd87afaabdb8bdc1  config-baseline.plugin.json
