+++ docs/.generated/plugin-sdk-api-baseline.sha256
@@ -1,2 +1,2 @@
-289d3f80e1f913a3205cf6b267dcf50f2a0a871a81d3ad3a00b7ae14339c2516  plugin-sdk-api-baseline.json
-731105cd51e816232c963618ca5051945f887455af8b178373bdd4ed57e36751  plugin-sdk-api-baseline.jsonl
+effb6ee16d16bc1b1e76ec293868910f887a168d9b756449928c703fe4c9e81a  plugin-sdk-api-baseline.json
+16eb8ac91b10b3ee62d856bf16c25c1ba3ba9fa7303500af2947a6e532b0c222  plugin-sdk-api-baseline.jsonl
