+++ extensions/qa-matrix/src/substrate/harness.runtime.test.ts
@@ -84,10 +84,6 @@ describe("matrix harness runtime", () => {
       expect(result.stopCommand).toBe(
         `docker compose -f ${outputDir}/docker-compose.matrix-qa.yml down --remove-orphans`,
       );
-      await result.restartService();
-      expect(calls).toContain(
-        `docker compose -f ${outputDir}/docker-compose.matrix-qa.yml restart matrix-qa-homeserver @/repo/openclaw`,
-      );
     } finally {
       await rm(outputDir, { recursive: true, force: true });
     }
