+++ extensions/telegram/setup-entry.ts
@@ -2,9 +2,6 @@ import { defineBundledChannelSetupEntry } from "openclaw/plugin-sdk/channel-entr
 
 export default defineBundledChannelSetupEntry({
   importMetaUrl: import.meta.url,
-  features: {
-    legacyStateMigrations: true,
-  },
   plugin: {
     specifier: "./setup-plugin-api.js",
     exportName: "telegramSetupPlugin",
