+++ extensions/telegram/src/channel.setup.ts
@@ -4,15 +4,10 @@ import type { TelegramProbe } from "./probe.js";
 import { telegramSetupAdapter } from "./setup-core.js";
 import { telegramSetupWizard } from "./setup-surface.js";
 import { createTelegramPluginBase } from "./shared.js";
-import { detectTelegramLegacyStateMigrations } from "./state-migrations.js";
 
 export const telegramSetupPlugin: ChannelPlugin<ResolvedTelegramAccount, TelegramProbe> = {
   ...createTelegramPluginBase({
     setupWizard: telegramSetupWizard,
     setup: telegramSetupAdapter,
   }),
-  lifecycle: {
-    detectLegacyStateMigrations: ({ cfg, env }) =>
-      detectTelegramLegacyStateMigrations({ cfg, env }),
-  },
 };
