+++ extensions/whatsapp/setup-entry.ts
@@ -2,10 +2,6 @@ import { defineBundledChannelSetupEntry } from "openclaw/plugin-sdk/channel-entr
 
 export default defineBundledChannelSetupEntry({
   importMetaUrl: import.meta.url,
-  features: {
-    legacyStateMigrations: true,
-    legacySessionSurfaces: true,
-  },
   plugin: {
     specifier: "./setup-plugin-api.js",
     exportName: "whatsappSetupPlugin",
