+++ extensions/whatsapp/src/channel.setup.ts
@@ -8,7 +8,6 @@ import {
 } from "./group-policy.js";
 import { whatsappSetupAdapter } from "./setup-core.js";
 import { createWhatsAppPluginBase, whatsappSetupWizardProxy } from "./shared.js";
-import { detectWhatsAppLegacyStateMigrations } from "./state-migrations.js";
 
 export const whatsappSetupPlugin: ChannelPlugin<ResolvedWhatsAppAccount> = {
   ...createWhatsAppPluginBase({
@@ -21,8 +20,4 @@ export const whatsappSetupPlugin: ChannelPlugin<ResolvedWhatsAppAccount> = {
     setup: whatsappSetupAdapter,
     isConfigured: async (account) => await webAuthExists(account.authDir),
   }),
-  lifecycle: {
-    detectLegacyStateMigrations: ({ oauthDir }) =>
-      detectWhatsAppLegacyStateMigrations({ oauthDir }),
-  },
 };
