+++ src/channels/plugins/bundled.ts
@@ -29,10 +29,6 @@ type BundledChannelSetupEntryRuntimeContract = {
   kind: "bundled-channel-setup-entry";
   loadSetupPlugin: () => ChannelPlugin;
   loadSetupSecrets?: () => ChannelPlugin["secrets"] | undefined;
-  features?: {
-    legacyStateMigrations?: boolean;
-    legacySessionSurfaces?: boolean;
-  };
 };
 
 type GeneratedBundledChannelEntry = {
@@ -90,13 +86,6 @@ function resolveChannelSetupModuleEntry(
     return null;
   }
   return record as BundledChannelSetupEntryRuntimeContract;
-}
-
-function hasSetupEntryFeature(
-  entry: BundledChannelSetupEntryRuntimeContract | undefined,
-  feature: keyof NonNullable<BundledChannelSetupEntryRuntimeContract["features"]>,
-): boolean {
-  return entry?.features?.[feature] === true;
 }
 
 function resolveBundledChannelBoundaryRoot(params: {
@@ -269,19 +258,6 @@ export function listBundledChannelPlugins(): readonly ChannelPlugin[] {
 
 export function listBundledChannelSetupPlugins(): readonly ChannelPlugin[] {
   return listBundledChannelPluginIds().flatMap((id) => {
-    const plugin = getBundledChannelSetupPlugin(id);
-    return plugin ? [plugin] : [];
-  });
-}
-
-export function listBundledChannelSetupPluginsByFeature(
-  feature: keyof NonNullable<BundledChannelSetupEntryRuntimeContract["features"]>,
-): readonly ChannelPlugin[] {
-  return listBundledChannelPluginIds().flatMap((id) => {
-    const setupEntry = getLazyGeneratedBundledChannelEntry(id, { includeSetup: true })?.setupEntry;
-    if (!hasSetupEntryFeature(setupEntry, feature)) {
-      return [];
-    }
     const plugin = getBundledChannelSetupPlugin(id);
     return plugin ? [plugin] : [];
   });
