+++ src/channels/plugins/session-conversation.ts
@@ -1,6 +1,5 @@
 import { getRuntimeConfigSnapshot } from "../../config/runtime-snapshot.js";
 import { tryLoadActivatedBundledPluginPublicSurfaceModuleSync } from "../../plugin-sdk/facade-runtime.js";
-import { getActivePluginChannelRegistryVersion } from "../../plugins/runtime.js";
 import {
   parseRawSessionConversationRef,
   parseThreadSessionSuffix,
@@ -58,16 +57,6 @@ type SessionConversationResolutionOptions = {
 type NormalizedSessionConversationResolution = ResolvedSessionConversation & {
   hasExplicitParentConversationCandidates: boolean;
 };
-
-type BundledSessionConversationFallbackCacheEntry = {
-  version: number;
-  resolveSessionConversation: BundledSessionKeyModule["resolveSessionConversation"] | null;
-};
-
-const bundledSessionConversationFallbackCache = new Map<
-  string,
-  BundledSessionConversationFallbackCacheEntry
->();
 
 function normalizeResolvedChannel(channel: string): string {
   return (
@@ -159,35 +148,22 @@ function resolveBundledSessionConversationFallback(params: {
     return null;
   }
   const dirName = normalizeResolvedChannel(params.channel);
-  const version = getActivePluginChannelRegistryVersion();
-  let cached = bundledSessionConversationFallbackCache.get(dirName);
-  if (!cached || cached.version !== version) {
-    let resolveSessionConversation: BundledSessionKeyModule["resolveSessionConversation"] | null =
-      null;
-    try {
-      const loaded = tryLoadActivatedBundledPluginPublicSurfaceModuleSync<BundledSessionKeyModule>({
+  let resolveSessionConversation: BundledSessionKeyModule["resolveSessionConversation"];
+  try {
+    resolveSessionConversation =
+      tryLoadActivatedBundledPluginPublicSurfaceModuleSync<BundledSessionKeyModule>({
         dirName,
         artifactBasename: SESSION_KEY_API_ARTIFACT_BASENAME,
-      });
-      resolveSessionConversation =
-        typeof loaded?.resolveSessionConversation === "function"
-          ? loaded.resolveSessionConversation
-          : null;
-    } catch {
-      resolveSessionConversation = null;
-    }
-    cached = {
-      version,
-      resolveSessionConversation,
-    };
-    bundledSessionConversationFallbackCache.set(dirName, cached);
+      })?.resolveSessionConversation;
+  } catch {
+    return null;
   }
-  if (typeof cached.resolveSessionConversation !== "function") {
+  if (typeof resolveSessionConversation !== "function") {
     return null;
   }
 
   return normalizeSessionConversationResolution(
-    cached.resolveSessionConversation({
+    resolveSessionConversation({
       kind: params.kind,
       rawId: params.rawId,
     }),
@@ -204,10 +180,6 @@ function isBundledSessionConversationFallbackDisabled(channel: string): boolean 
   }
   const entry = snapshot.plugins.entries?.[normalizeResolvedChannel(channel)];
   return !!entry && typeof entry === "object" && entry.enabled === false;
-}
-
-function shouldProbeBundledSessionConversationFallback(rawId: string): boolean {
-  return rawId.includes(":");
 }
 
 function resolveSessionConversationResolution(params: {
@@ -228,10 +200,7 @@ function resolveSessionConversationResolution(params: {
       rawId,
     }),
   );
-  const shouldTryBundledFallback =
-    params.bundledFallback !== false &&
-    !messaging &&
-    shouldProbeBundledSessionConversationFallback(rawId);
+  const shouldTryBundledFallback = params.bundledFallback !== false && !messaging;
   const resolved =
     pluginResolved ??
     (shouldTryBundledFallback
