+++ src/commands/agents.bind.test-support.ts
@@ -39,16 +39,10 @@ vi.mock("../config/config.js", async () => {
 export const runtime = createTestRuntime();
 
 let agentsCommandModulePromise: Promise<typeof import("./agents.js")> | undefined;
-let agentsBindCommandModulePromise: Promise<typeof import("./agents.commands.bind.js")> | undefined;
 
 export async function loadFreshAgentsCommandModuleForTest() {
   agentsCommandModulePromise ??= import("./agents.js");
   return await agentsCommandModulePromise;
-}
-
-export async function loadFreshAgentsBindCommandModuleForTest() {
-  agentsBindCommandModulePromise ??= import("./agents.commands.bind.js");
-  return await agentsBindCommandModulePromise;
 }
 
 export function resetAgentsBindTestHarness(): void {
