+++ src/commands/sessions-cleanup.ts
@@ -20,14 +20,12 @@ import {
   type SessionStoreTarget,
 } from "./session-store-targets.js";
 import {
-  resolveSessionDisplayDefaults,
-  resolveSessionDisplayModel,
-} from "./sessions-display-model.js";
-import {
   formatSessionAgeCell,
   formatSessionFlagsCell,
   formatSessionKeyCell,
   formatSessionModelCell,
+  resolveSessionDisplayDefaults,
+  resolveSessionDisplayModel,
   SESSION_AGE_PAD,
   SESSION_KEY_PAD,
   SESSION_MODEL_PAD,
@@ -282,7 +280,7 @@ function renderStoreDryRunPlan(params: {
   ].join(" ");
   params.runtime.log(rich ? theme.heading(header) : header);
   for (const actionRow of params.actionRows) {
-    const model = resolveSessionDisplayModel(params.cfg, actionRow);
+    const model = resolveSessionDisplayModel(params.cfg, actionRow, params.displayDefaults);
     const line = [
       formatCleanupActionCell(actionRow.action, rich),
       formatSessionKeyCell(actionRow.key, rich),
