+++ src/gateway/startup-auth.test.ts
@@ -1,11 +1,6 @@
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import type { OpenClawConfig } from "../config/config.js";
 import { expectGeneratedTokenPersistedToGatewayAuth } from "../test-utils/auth-token-assertions.js";
-import {
-  assertGatewayAuthNotKnownWeak,
-  assertHooksTokenSeparateFromGatewayAuth,
-  ensureGatewayStartupAuth,
-} from "./startup-auth.js";
 
 const mocks = vi.hoisted(() => ({
   replaceConfigFile: vi.fn(async (_params: { nextConfig: OpenClawConfig }) => {}),
@@ -18,6 +13,19 @@ vi.mock("../config/config.js", async () => {
     replaceConfigFile: mocks.replaceConfigFile,
   };
 });
+
+let assertGatewayAuthNotKnownWeak: typeof import("./startup-auth.js").assertGatewayAuthNotKnownWeak;
+let assertHooksTokenSeparateFromGatewayAuth: typeof import("./startup-auth.js").assertHooksTokenSeparateFromGatewayAuth;
+let ensureGatewayStartupAuth: typeof import("./startup-auth.js").ensureGatewayStartupAuth;
+
+async function loadFreshStartupAuthModuleForTest() {
+  vi.resetModules();
+  ({
+    assertGatewayAuthNotKnownWeak,
+    assertHooksTokenSeparateFromGatewayAuth,
+    ensureGatewayStartupAuth,
+  } = await import("./startup-auth.js"));
+}
 
 describe("ensureGatewayStartupAuth", () => {
   async function expectEphemeralGeneratedTokenWhenOverridden(cfg: OpenClawConfig) {
@@ -35,9 +43,10 @@ describe("ensureGatewayStartupAuth", () => {
     expect(mocks.replaceConfigFile).not.toHaveBeenCalled();
   }
 
-  beforeEach(() => {
+  beforeEach(async () => {
     vi.restoreAllMocks();
     mocks.replaceConfigFile.mockClear();
+    await loadFreshStartupAuthModuleForTest();
   });
 
   async function expectNoTokenGeneration(cfg: OpenClawConfig, mode: string) {
@@ -467,9 +476,8 @@ describe("ensureGatewayStartupAuth", () => {
 });
 
 describe("assertGatewayAuthNotKnownWeak", () => {
-  beforeEach(() => {
-    vi.restoreAllMocks();
-    mocks.replaceConfigFile.mockClear();
+  beforeEach(async () => {
+    await loadFreshStartupAuthModuleForTest();
   });
 
   it("throws on the known-weak token sentinel", () => {
