+++ src/infra/exec-approval-forwarder.ts
@@ -1,8 +1,5 @@
 import type { ReplyPayload } from "../auto-reply/types.js";
-import {
-  getLoadedChannelPlugin,
-  resolveChannelApprovalAdapter,
-} from "../channels/plugins/index.js";
+import { getChannelPlugin, resolveChannelApprovalAdapter } from "../channels/plugins/index.js";
 import { loadConfig } from "../config/config.js";
 import type {
   ExecApprovalForwardingConfig,
@@ -201,7 +198,7 @@ function shouldSkipForwardingFallback(params: {
   if (!channel) {
     return false;
   }
-  const adapter = resolveChannelApprovalAdapter(getLoadedChannelPlugin(channel));
+  const adapter = resolveChannelApprovalAdapter(getChannelPlugin(channel));
   return (
     adapter?.delivery?.shouldSuppressForwardingFallback?.({
       cfg: params.cfg,
@@ -380,7 +377,7 @@ function buildApprovalRenderPayload<TParams>(params: {
 }): ReplyPayload {
   const channel = normalizeMessageChannel(params.target.channel) ?? params.target.channel;
   const adapterPayload = channel
-    ? params.resolveRenderer(resolveChannelApprovalAdapter(getLoadedChannelPlugin(channel)))?.(
+    ? params.resolveRenderer(resolveChannelApprovalAdapter(getChannelPlugin(channel)))?.(
         params.renderParams,
       )
     : null;
@@ -586,7 +583,7 @@ function createApprovalHandlers<
         if (!channel) {
           return;
         }
-        await getLoadedChannelPlugin(channel)?.outbound?.beforeDeliverPayload?.({
+        await getChannelPlugin(channel)?.outbound?.beforeDeliverPayload?.({
           cfg,
           target,
           payload,
