+++ src/infra/restart-stale-pids.test.ts
@@ -822,8 +822,6 @@ describe.skipIf(isWindows)("restart-stale-pids", () => {
       const stalePid = process.pid + 912;
       Object.defineProperty(process, "platform", { value: "win32", configurable: true });
       try {
-        let fakeNow = 0;
-        __testing.setDateNowOverride(() => fakeNow);
         mockReadWindowsListeningPidsResult.mockReturnValue({ ok: true, pids: [stalePid] });
         mockReadWindowsProcessArgs.mockReturnValue(["openclaw", "gateway"]);
         mockReadWindowsProcessArgsResult.mockReturnValue({
@@ -849,9 +847,6 @@ describe.skipIf(isWindows)("restart-stale-pids", () => {
           }
           return true;
         });
-        __testing.setSleepSyncOverride((ms) => {
-          fakeNow += ms;
-        });
 
         expect(cleanStaleGatewayProcessesSync()).toEqual([]);
         expect(mockSpawnSync).toHaveBeenNthCalledWith(
@@ -867,8 +862,6 @@ describe.skipIf(isWindows)("restart-stale-pids", () => {
           expect.objectContaining({ timeout: 5000 }),
         );
       } finally {
-        __testing.setSleepSyncOverride(null);
-        __testing.setDateNowOverride(null);
         if (origDescriptor) {
           Object.defineProperty(process, "platform", origDescriptor);
         }
